Statistics/Challenges minor updates

+++ Game.Abstractions/Entities/PlayerChallenge.cs
@@ -4,7 +4,7 @@ namespace Game.Abstractions.Entities
     {
         public int PlayerId { get; set; }
         public int ChallengeId { get; set; }
-        public int Progress { get; set; }
+        public decimal Progress { get; set; }
         public bool Completed { get; set; }
         public DateTime? CompletedAt { get; set; }
