+++ Game.Core/Battle/BattleStats.cs
@@ -2,9 +2,9 @@ namespace Game.Core.Battle
 {
     public class BattleStats
     {
-        public long PlayerDamageDealt { get; set; }
-        public long PlayerDamageTaken { get; set; }
-        public long HighestPlayerAttack { get; set; }
+        public double PlayerDamageDealt { get; set; }
+        public double PlayerDamageTaken { get; set; }
+        public double HighestPlayerAttack { get; set; }
         public int PlayerSkillsUsed { get; set; }
     }
 }
